+++ internal/modules/warehouse/domain/entities/product/product_repository.go
@@ -6,8 +6,8 @@ type Repository interface {
 	GetPaginated(ctx context.Context, limit, offset int, sortBy []string) ([]*Product, error)
 	Count(ctx context.Context) (int64, error)
 	GetAll(ctx context.Context) ([]*Product, error)
-	GetByID(ctx context.Context, id int64) (*Product, error)
+	GetByID(ctx context.Context, id uint) (*Product, error)
 	Create(ctx context.Context, data *Product) error
 	Update(ctx context.Context, data *Product) error
-	Delete(ctx context.Context, id int64) error
+	Delete(ctx context.Context, id uint) error
 }
